requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
Click==7.0
cryptography==2.1.4
Flask==1.0.2
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.1.0
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.7.1
python-dbusmock==0.17.1
python-dotenv==0.10.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
Werkzeug==0.14.1
